requirements.txt
fastapi
sqlalchemy
uvicorn
pydantic
dotenv
bcrypt
asyncpg
pydantic[email]
passlib
python-jose
python-multipart
aiosmtplib
